% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chorddiagrams_tikka51124.R
\name{group_chords}
\alias{group_chords}
\title{Chord diagram}
\usage{
group_chords(
  vars,
  n_level,
  fig_name,
  big,
  rem,
  modi,
  colt,
  gend,
  colors,
  a,
  b,
  c,
  d,
  e,
  f
)
}
\arguments{
\item{vars}{The dataset with steroids and cases}

\item{n_level}{The case (e.g. steatosis). Typically a column name in the dataset}

\item{fig_name}{An header add (of the case) for the forestplot's header}

\item{gend}{All, Female, Male (typically a gender column)}

\item{oute}{The label title. Here, almost the same as case.}
}
\value{
The forestplots. Yes.
}
\description{
Converts data to forestplots.
Written by Pauli Tikka. University of Turku. Updated 4.11.24.
}
\examples{
Outcome='Steatosis.Grade.0.To.3';Out='Steatosis'; oute='Steatosis';first=TRUE; e='P4';ordera=c();
Group='All';name1=paste("Forest plot of",Group, "Steroid Ratios in",Out);ordera=forestplots(NAFLD,Outcome,Group,name1,ordera,oute,first,e,xlim)
Afterwards: first=FALSE;
}
